+++ app/Dtos/LikeDto.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Dtos;
+
+class LikeDto {
+
+    //code
+
+}