+++ app/Http/Controllers/HomeController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Question;
 use Illuminate\Http\Request;
 
 class HomeController extends Controller
@@ -9,8 +10,8 @@ class HomeController extends Controller
 
     public function index()
     {
+        $questions = Question::limit(10)->orderBy('created_at', 'desc')->get();
 
-        $questions = [];
         return view('welcome', compact('questions'));
     }