+++ app/Http/Controllers/LikeController.php
@@ -1,0 +1,79 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use App\Http\Requests\LikeRequest;
+use App\Models\Answer;
+use App\Models\Like;
+use App\Models\Question;
+use Illuminate\Http\Request;
+
+class LikeController extends Controller
+{
+
+    public function like(Request $request) {
+
+        $likeableType = $this->getLikeAbleType($request->likeable_type);
+
+        $liked = Like::updateOrCreate(
+            [
+               'likeable_type' => $likeableType,
+               'likeable_id' => $request->likeable_id,
+               'user_id' => \Auth::id()
+            ],
+            [
+                'like' => true
+            ]
+        );
+
+        if($liked) return back();
+
+        return back();
+
+    }
+
+    public function dislike(Request $request)
+    {
+        $likeableType = $this->getLikeAbleType($request->likeable_type);
+
+        $disliked = Like::updateOrCreate(
+            [
+                'likeable_type' => $likeableType,
+                'likeable_id' => $request->likeable_id,
+                'user_id' => \Auth::id()
+            ],
+            [
+                'like' => false
+            ]
+        );
+
+        if($disliked) return back();
+
+        return back();
+
+    }
+
+
+
+    public function destroy($id)
+    {
+        $like = Like::destroy($id);
+        if($like) return back();
+
+        return back();
+    }
+
+
+    public function getLikeAbleType($likeable_type) : string
+    {
+        switch ($likeable_type){
+            case 'question':
+                $likeable_type = Question::class;
+                break;
+            case 'answer':
+                $likeable_type  = Answer::class;
+                break;
+        }
+        return $likeable_type;
+    }
+}